Initial stab at textbox support.
This is a basic async text input system. When a textbox is active, all
input is redirected to it, until it is marked as inactive. A callback
handles processing whatever text was put in it; that callback should
probably handle cleaning up the contents of the textbox, too, but right
now it auto-clears.

It also horks on non-printing characters, treating most things like
Return.